requirements.txt
asgiref==3.8.1
autopep8==2.1.0
certifi==2023.11.17
distlib==0.3.8
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
filelock==3.13.1
greenlet==3.0.3
numpy==1.26.2
pillow==10.3.0
pipenv==2023.12.1
platformdirs==4.2.0
pycodestyle==2.11.1
PyJWT==1.7.1
setuptools==69.0.3
SQLAlchemy==2.0.27
sqlparse==0.5.0
typing==3.7.4.3
typing_extensions==4.9.0
virtualenv==20.25.1
